Refactor booking and tent hooks to remove unnecessary console logs and improve error handling; delete unused overview, revenue breakdown, total guest, and refunds hooks; reintroduce overview and revenue breakdown hooks with improved structure; enhance total guest and refunds hooks for better data management and error reporting.
