+++ src/app/(dashboard)/tent-management/tent-categories/page.tsx
@@ -19,7 +19,6 @@ export default function TentCategoriesPage() {
 	} = useCategory();
 
 	useEffect(() => {
-		console.log('Fetching categories data...');
 		getCategories();
 	}, []);
